# Conflicts:
#	src/main/java/lk/gov/mohe/adminsystem/permission/Permission.java
#	src/main/java/lk/gov/mohe/adminsystem/permission/PermissionCategory.java
#	src/main/java/lk/gov/mohe/adminsystem/permission/PermissionController.java
#	src/main/java/lk/gov/mohe/adminsystem/permission/PermissionRepository.java
#	src/main/java/lk/gov/mohe/adminsystem/role/RoleController.java
#	src/main/java/lk/gov/mohe/adminsystem/role/RoleDto.java

+++ src/main/java/lk/gov/mohe/adminsystem/user/UserDto.java
@@ -4,13 +4,11 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record UserDto(
-    Long id,
+    Integer id,
     String username,
     String email,
     String fullName,
     String phoneNumber,
     String role,
     String division,
-    Boolean isActive
-) {
-}
+    Boolean isActive) {}
